requirements.txt
asyncio==3.4.3
Flask==1.1.2
flask-oidc==1.4.0
Jinja2==2.11.2
python-dotenv==0.15.0
requests==2.25.1
WTForms==2.3.3